requirements.txt
openai>=2.26.0
openai-agents>=0.13.0,<1.0.0
streamlit>=1.46.0
python-dotenv>=1.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
pydantic>=2.0.0

